requirements.txt
streamlit
PyMuPDF
langchain
langchain-google-genai
google-generativeai
python-dotenv